// AppCovid-19/src/models/usuario-registro.model.ts
export class UsuarioRegistroModel {
    nombre: string;
    email: string;
    codigoPostal: number;
    fechaNacimiento: string;
    password: string;
    instancia: number;
    idEncuesta: number;
    idTicket: number;
    ubicacion: string;
    Telefono: string;
    Calle: string;
    Numero: string;
    Municipio: string;
    idGenero: number;
}